# Conflicts:
#	스프링부트/src/main/resources/templates/js/ui_basic.js

+++ 스프링부트/src/main/java/party/people/repository/place/PlaceInterface.java
@@ -30,6 +30,6 @@ public interface PlaceInterface {
 
     List<Place>randon3();
 
-    Place idSearch(int id);
+    Place idSearch(Long id);
 
 }
