+++ 스프링부트/src/main/java/party/people/repository/place/PlaceMapper.java
@@ -32,6 +32,6 @@ public interface PlaceMapper {
 
     List<Place>randon3();
 
-    Place idSearch(int id);
+    Place idSearch(Long id);
 
 }
